+++ CHANGES.md
@@ -1,4 +1,5 @@
 #6.0.9 2018-??-??
+ - 2018-06-12 Fixed bug#[13902](https://bugs.otrs.org/show_bug.cgi?id=13902) - Dialog submit is not possible for an appointment created by rule based on pending time.
  - 2018-06-12 Fixed bug#[13888](https://bugs.otrs.org/show_bug.cgi?id=13888) - Wrong charset in output while upgrading packages.
  - 2018-06-07 Enhanced package manager:
    - Not verified packages can't be installed by default (via GUI + OTRS console).